Splitting out schema maps to separate files + access via functions
